Fixed Arithmetic Overflow by casting numElements to size_t

+++ Plugins/bullet3/src/LinearMath/btSerializer.h
@@ -608,7 +608,7 @@ public:
 
 	virtual btChunk* allocate(size_t size, int numElements)
 	{
-		unsigned char* ptr = internalAlloc(int(size) * numElements + sizeof(btChunk));
+		unsigned char* ptr = internalAlloc(int(size) * (size_t)numElements + sizeof(btChunk));
 
 		unsigned char* data = ptr + sizeof(btChunk);
